backend/requirements.txt
fastapi
uvicorn
pdfplumber
google-generativeai
pydantic
python-dotenv
python-multipart
